requirements.txt
filelock==3.0.12
numba>=0.47.0,<0.49.0
numpy==1.18.1
pandas==1.0.1
sacred==0.8.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
seml==0.2.4
tabulate==0.8.7
tinydb==4.2.0
tinydb-serialization==2.0.0
torchvision
tqdm==4.42.1